+++ tests/production/e2e/07-idempotency-handling.test.ts
@@ -95,7 +95,7 @@ describe('PRODUCTION E2E: Idempotency Handling', function () {
     const requestBody = {
       nftMint: testNft.mint.toString(),
       swapType: 'NFT_FOR_SOL',
-      solAmount: parseInt(PRODUCTION_CONFIG.swapAmount) * 1_000_000_000, // Convert SOL to lamports
+      solAmount: parseFloat(PRODUCTION_CONFIG.swapAmount) * 1_000_000_000, // Convert SOL to lamports
       seller: wallets.sender.publicKey.toString(),
       buyer: wallets.receiver.publicKey.toString(),
       expiry: expiry.toISOString(), // Same expiry for both
